// src/types/user.ts
export type UserType = {
  address?: string;
  companyName?: string;
  email?: string;
  firstName?: string;
  id?: string;
  isAdmin: boolean;
  isBillingContact?: string;
  lastName?: string;
  mainAvatarUrl?: string;
  organizationInfo?: string;
  phoneNumber?: string;
  thumbnailAvatarUrl?: string;
  username?: string;
};